Use 32 bit values for stream slots

+++ libcaf_core/caf/fwd.hpp
@@ -198,7 +198,7 @@ using ip_endpoint = ipv6_endpoint;
 using ip_subnet = ipv6_subnet;
 using settings = dictionary<config_value>;
 using skippable_result = variant<delegated<message>, message, error, skip_t>;
-using stream_slot = uint16_t;
+using stream_slot = uint32_t;
 using type_id_t = uint16_t;
 
 // -- functions ----------------------------------------------------------------
